tools/requirements.txt
# PCA Reduction Tool Dependencies
numpy>=1.24.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0
